requirements.txt
amqp==5.0.1
billiard==3.6.3.0
celery==5.0.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Flask==1.1.2
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
six==1.15.0
vine==5.0.0
waitress==1.4.4
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0
